requirements.txt
streamlit
streamlit-image-comparison>=0.0.4
opencv-python-headless>=4.8
numpy>=1.26
Pillow>=9.0